% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1stfunction.R
\name{AdjustTiter}
\alias{AdjustTiter}
\title{Adjusting Titers}
\usage{
AdjustTiter(df, col_vec, min_thres = NA, val_min = log2(10),
  max_thres = NA, val_max = log2(1280))
}
\arguments{
\item{col_vec}{Columns that indicate titers}

\item{min_thres}{minimum threshold}

\item{max_thres}{maximum threshold}
}
\value{
Titers which have no value below log2(10) or above log2(1280)
}
\description{
This function is to adjust titers to the limit of detection from log2(10) to log2(1280)
}
