% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1stfunction.R
\name{AssignSerotypeHomo}
\alias{AssignSerotypeHomo}
\title{Assigning serotype of homotypic infection samples.}
\usage{
AssignSerotypeHomo(df, col_vec, cutoff = 5)
}
\arguments{
\item{df}{data.frame}

\item{col_vec}{vector of Ags variables}

\item{cutoff}{default is 5}
}
\value{
df with an extra col called pred.serotype
}
\description{
Which antigen has titer that is greater than cutoff ( default = 5)
is assigned such antigen as an infecting serotype
}
