% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1stfunction.R
\name{createAgArraySet.new}
\alias{createAgArraySet.new}
\title{Creating an agArraySet object}
\usage{
createAgArraySet.new(metadata, agArrayList, Value = "Median",
  Func = robustComputeTiter.new, parallel = TRUE, pos.contr.label = NULL,
  exclude.trays = NULL, exclude.antigens = NULL)
}
\value{
An agArraySet object
}
\description{
This function is edited from the function "createAgArraySet", package "aatk".agArrayList is kept as original regardless of Flag label.
}
