% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1stfunction.R
\name{E_Status}
\alias{E_Status}
\title{Transforming ELISA_PanbioUnit into ELISA Status c("Neg","Pos","Uncertain")}
\usage{
E_Status(df, var)
}
\arguments{
\item{df}{data.frame}

\item{var}{a vector of ELISA result in PanbioUnit format}
}
\value{
df with an extra col called " E_Status"
}
\description{
Transforming ELISA_PanbioUnit into ELISA Status c("Neg","Pos","Uncertain")
}
