% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1stfunction.R
\name{PMAStatus}
\alias{PMAStatus}
\title{Assigning PMA status}
\usage{
PMAStatus(df, col_vec, cutoff = 5)
}
\arguments{
\item{df}{data.frame}

\item{col_vec}{vector of Ags variables}

\item{cutoff}{default is 5}
}
\value{
df with an extra col called "cutoff_5"
}
\description{
Assigning PMA status
}
