% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1stfunction.R
\name{robustComputeTiter.new}
\alias{robustComputeTiter.new}
\title{Titer computtation}
\usage{
robustComputeTiter.new(dilutionvalues, log2.transform = FALSE, top = 65535,
  bottom = 3000, monotonicity.cutoff = 0.5, log2.output = TRUE,
  generate.plot = FALSE)
}
\value{
EC50 value; if luminescence at dilution 20 < top*0.5, EC50 is assigned into 2 groups:10 or 20.
}
\description{
This is a function that performs titer computation for an antigen, edited from "robustComputeTiter{aatk"}
}
