% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1stfunction.R
\name{tong}
\alias{tong}
\title{Description of the method}
\usage{
tong(a, b)
}
\arguments{
\item{a}{First number}

\item{b}{second number}
}
\value{
sum of a and b
}
\description{
Description of the method
}
